Complete Implementation of removeStacksById function
Include a conditional to check for multiple items to delete

+++ controllers/stackControllers.js
@@ -12,15 +12,21 @@ async function getStacks(req, res, next){
 }
 
 async function removeStacksById(req, res, next){
-    console.log('REMOVING STACKS')
-    const {usedCodes} = req.body
-    console.log(req.body)
-    console.log(usedCodes)
-  
-  const removedStack = await StackModel.deleteOne({stackId:usedCodes[0]})
-
-  console.log(removedStack)
-  res.status(200).send({removedStack})
+  const {usedCodes} = req.body 
+  let deleteReport
+  if (usedCodes.length === 1) {
+    deleteReport = 
+    await StackModel.deleteOne({stackId:usedCodes[0]})
+  }
+  if (usedCodes.length > 1) {
+    deleteReport = 
+    await StackModel.deleteMany({
+      stackId: {
+        $in: usedCodes
+      }
+    })
+  }
+  res.status(200).send({deleteReport})
 }
 
 module.exports = {addStack, getStacks, removeStacksById}
